// web/apps/msteams/src/components/organization/organizationTeams/__generated__/teamCard_TeamDetails.graphql.ts
export type teamCard_TeamDetails$data = {
  readonly canDelete: boolean;
  readonly canModify: boolean;
  readonly hasFutureBooking: boolean;
  readonly id: string;
  readonly members: ReadonlyArray<{
    readonly organizationMember: {
      readonly customer: {
        readonly familyName: string | null | undefined;
        readonly givenName: string | null | undefined;
        readonly middleName: string | null | undefined;
        readonly name: string | null | undefined;
        readonly photoUrl: string | null | undefined;
        readonly uniqueId: string;
      };
      readonly uniqueId: string;
    } | null | undefined;
  }>;
  readonly name: string;
  readonly organization: {
    readonly uniqueId: string;
  } | null | undefined;
  readonly " $fragmentType": "teamCard_TeamDetails";
};
export type teamCard_TeamDetails$key = {
  readonly " $data"?: teamCard_TeamDetails$data;
  readonly " $fragmentSpreads": FragmentRefs<"teamCard_TeamDetails">;
};